adds String allRatings() extention

+++ src/main/java/IceCreamStall.java
@@ -1,6 +1,6 @@
 public class IceCreamStall extends Stall {
 
-    public IceCreamStall(String name, String ownerName, String parkingSpot, int rating) {
+    public IceCreamStall(String name, String ownerName, String parkingSpot, double rating) {
         super(name, ownerName, parkingSpot, rating);
     }
